Renamed some components for clarity
