+++ src/pages/AppHeader.vue
@@ -67,7 +67,7 @@ import ContentDivider from '@/components/contentAlignment/ContentDivider.vue'
 import { useRouter } from 'vue-router'
 import { useAuth } from '@/auth/authStores/auth'
 import { keycloak } from '@/auth/keycloak'
-import { useAccommodationStore } from '@/stores/AccomodationStore'
+import { useAccommodationStore } from '@/stores/AccommodationStore'
 import { ref } from 'vue'
 import { onClickOutside } from '@vueuse/core'
