+++ src/pages/TableView/TableRows.vue
@@ -67,7 +67,7 @@ SPDX-License-Identifier: AGPL-3.0-or-later
   import DetailButton from '@/components/buttons/DetailButton.vue'
 
   import { ref } from 'vue'
-  import { useAccommodationStore } from '@/stores/AccomodationStore'
+  import { useAccommodationStore } from '@/stores/AccommodationStore'
   import ImagesCarousel from './images/ImagesCarousel.vue'
   import type { Accommodation } from './types';
   import type { ColumnItem } from './ColumnValues'
