+++ src/pages/TableView/images/showImages.vue
@@ -45,7 +45,7 @@ SPDX-License-Identifier: AGPL-3.0-or-later
  
   import type { Accommodation } from '../types';
   import { computed, ref } from 'vue'
-  import { useAccommodationStore } from '@/stores/AccomodationStore';
+  import { useAccommodationStore } from '@/stores/AccommodationStore';
   
   import PlaceHolderImage from '@/components/Image/PlaceHolderImage.vue';
   import ShowImageFullView from '@/components/Image/showImageFullView.vue';
