+++ src/pages/TableView/sortPopup.vue
@@ -41,8 +41,8 @@ SPDX-License-Identifier: AGPL-3.0-or-later
 <script setup lang="ts">
     import { ChevronDoubleUpIcon, ChevronDoubleDownIcon } from '@heroicons/vue/24/outline';
 
-    import DatasetHeaderButton from '@/components/buttons/datasetHeaderButton.vue';
-    import { useAccommodationStore } from '@/stores/AccomodationStore';
+    import DatasetHeaderButton from '@/components/buttons/Button.vue';
+    import { useAccommodationStore } from '@/stores/AccommodationStore';
     import { useRoute, useRouter } from 'vue-router';
     import { onMounted, ref } from 'vue';
