+++ src/pages/datasetHeader/DatasetHeader.vue
@@ -108,7 +108,7 @@ SPDX-License-Identifier: AGPL-3.0-or-later
   import { InformationCircleIcon, Bars3BottomRightIcon, GlobeEuropeAfricaIcon, CursorArrowRaysIcon, SparklesIcon } from '@heroicons/vue/24/outline';  
 
 
-  import DatasetHeaderButton from '@/components/buttons/datasetHeaderButton.vue';
+  import DatasetHeaderButton from '@/components/buttons/Button.vue';
   import DatasetHeaderDropDown from '@/components/datasetHeader/DropDownMenu.vue';
   import DatasetHeaderSearchBar from '@/components/datasetHeader/SearchBar.vue';
   import InfoCard from './InfoCard.vue';
@@ -117,7 +117,7 @@ SPDX-License-Identifier: AGPL-3.0-or-later
   import ContentDivider from '@/components/contentAlignment/ContentDivider.vue';
 
   import { ref, watch } from 'vue';
-  import { useAccommodationStore } from '@/stores/AccomodationStore';
+  import { useAccommodationStore } from '@/stores/AccommodationStore';
   import { useRoute, useRouter } from 'vue-router';
   import { onClickOutside } from '@vueuse/core';
