+++ src/pages/filter/filterTable.vue
@@ -107,7 +107,7 @@ SPDX-License-Identifier: AGPL-3.0-or-later
   import { ref, computed, watch } from 'vue';
 
 
-  import { useAccommodationStore } from '@/stores/AccomodationStore';
+  import { useAccommodationStore } from '@/stores/AccommodationStore';
 
   import { useRoute, useRouter } from 'vue-router';
